feat: functioning one-way editor client
Squashed commit of the following:

commit 4324afd89c49814eb4943e758ddc508cd005c5cd
Author: Adam Zhang <898544@lcps.org>
Date:   Mon Sep 16 10:47:55 2024 -0400

    feat: functioning one-way editor client

    Right now, editor is one-way, but that's 50% of the way there 🥳.

commit 45d477abf35e07f85092f1027b7906b298953679
Author: Adam Zhang <898544@lcps.org>
Date:   Mon Sep 16 10:35:50 2024 -0400

    refactor: move Change enum to client module

commit 68681c855beec184936e9d1ca2c959638fa81c0f
Author: Adam Zhang <898544@lcps.org>
Date:   Mon Sep 16 10:34:56 2024 -0400

    feat(main): proof-of-concept for deserializing changes

commit 5f48431c1ab7e839f0724fa2aa968df4b81f65ca
Author: Adam Zhang <898544@lcps.org>
Date:   Mon Sep 16 10:23:04 2024 -0400

    feat(main): proof-of-concept for async stdin read

+++ src/main.rs
@@ -8,54 +8,64 @@ async fn main() {
     let stream: TcpStream;
 
     let args = std::env::args().collect::<Vec<String>>();
-    let server = args.contains(&String::from("server"));
-    if server {
-        let listener = TcpListener::bind("0.0.0.0:6969").await.unwrap();
-        stream = listener.accept().await.unwrap().0;
-    } else {
+    let cli = args.contains(&String::from("cli"));
+    if cli {
         let mut input = String::new();
         println!("Enter server IP: ");
         std::io::stdin().read_line(&mut input).unwrap();
         let addr = format!("{}:6969", input.trim());
         stream = TcpStream::connect(&addr).await.unwrap();
+    } else {
+        let listener = TcpListener::bind("0.0.0.0:6969").await.unwrap();
+        stream = listener.accept().await.unwrap().0;
     }
+
+    eprintln!("Connected");
 
     let mut client = Client::new(stream);
 
     client.begin_update_task();
 
-    let mut input = String::new();
-    loop {
-        println!("Current text: {:?}", client.read().await);
+    if cli {
+        let mut input = String::new();
+        loop {
+            println!("Current text: {:?}", client.read().await);
 
-        input.clear();
-        std::io::stdin().read_line(&mut input).unwrap();
+            input.clear();
+            std::io::stdin().read_line(&mut input).unwrap();
 
-        if input.trim().is_empty() {
-            continue;
-        }
-
-        let (command, rest) = input.split_at(2);
-        match command {
-            "i " => {
-                let (index, text) = rest.split_at(rest.find(' ').unwrap());
-                let index = index.trim().parse::<usize>().unwrap();
-                let text = text.trim();
-                client.insert_string(index, text).await;
-            }
-            "d " => {
-                let (begin, len) = rest.split_at(rest.find(' ').unwrap());
-                let begin = begin.trim().parse::<usize>().unwrap();
-                let len = len.trim().parse::<usize>().unwrap();
-                client.delete_string(begin, len).await;
-            }
-            _ => {
-                eprintln!("Invalid command");
+            if input.trim().is_empty() {
                 continue;
             }
-        }
 
-        eprintln!("Broadcasting changes");
-        client.broadcast_changes().await;
+            let (command, rest) = input.split_at(2);
+            match command {
+                "i " => {
+                    let (index, text) = rest.split_at(rest.find(' ').unwrap());
+                    let index = index.trim().parse::<usize>().unwrap();
+                    let text = text.trim();
+                    client.insert_string(index, text).await;
+                }
+                "d " => {
+                    let (begin, len) = rest.split_at(rest.find(' ').unwrap());
+                    let begin = begin.trim().parse::<usize>().unwrap();
+                    let len = len.trim().parse::<usize>().unwrap();
+                    client.delete_string(begin, len).await;
+                }
+                _ => {
+                    eprintln!("Invalid command");
+                    continue;
+                }
+            }
+
+            eprintln!("Broadcasting changes");
+            client.broadcast_changes().await;
+        }
+    } else {
+        client.begin_stdin_task();
+        loop {
+            tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+            client.broadcast_changes().await;
+        }
     }
 }
